+++ ccxt.js
@@ -7362,6 +7362,10 @@ var coinfloor = {
     parseTicker (ticker, market) {
         // rewrite to get the timestamp from HTTP headers
         let timestamp = this.milliseconds ();
+        // they sometimes return null for vwap
+        let vwap = undefined;
+        if (('vwap' in ticker) && (ticker['vwap']))
+            vwap = parseFloat (ticker['vwap']);
         return {
             'timestamp': timestamp,
             'datetime': this.iso8601 (timestamp),
@@ -7369,7 +7373,7 @@ var coinfloor = {
             'low': parseFloat (ticker['low']),
             'bid': parseFloat (ticker['bid']),
             'ask': parseFloat (ticker['ask']),
-            'vwap': parseFloat (ticker['vwap']),
+            'vwap': vwap,
             'open': undefined,
             'close': undefined,
             'first': undefined,